refactor(repository): JobPostingRepository 헬퍼/정산 모듈 분할 (Phase 1)
- JobPostingRepositoryHelpers.ts: 테이블 상수·toJobPosting·검증/접근 헬퍼·buildSlotRoleKey 추출
- JobPostingRepositorySettlement.ts: 정산 역할 키/병합 순수 헬퍼 추출
- 진입점에 buildSlotRoleKey re-export 유지(테스트 경로 직접 import 호환)
- 클래스 본문·시그니처 불변, 코드 이동만 (921→699줄, tsc 0, test 4286 pass)
